% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soc_find_team_stat.R
\name{soc_find_team_stat}
\alias{soc_find_team_stat}
\title{Find team statistic}
\usage{
soc_find_team_stat(df, team_name, feature = age)
}
\arguments{
\item{df}{The dataframe of the soccer dataset}

\item{team_name}{String in the dataframe. Users can input any string from the "Team" column.}

\item{feature}{Optional argument. The default feature is "age". Users can input other numerical features.}
}
\value{
list A list contains a descriptive statistics table and a box plot to show a certain descriptive statistic feature of a particular team.
}
\description{
Returns the descriptive statistic table and box plot of a particular team.
This function can help users to find the statistical data of a particular team.
A box plot is provided for visualization of the statistical data.
Users can visualize a certain feature of the team by inputting it in the optional argument "feature".
If no input is found, the default feature of the box plot will be the "age" of the user's inputted team.
}
\examples{
soc_find_team_stat(soccer_df, "Manchester United", "Market_Value_Euros")
}
