% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_identification.R
\name{soc_get_oultiers}
\alias{soc_get_oultiers}
\title{Returns outliers in the dataset based on values of a variable
This function identifies outliers in the dataset based on either of the following methods:
\enumerate{
\item Interquartile Range (IQR) Method: Identifies all values less than Q1 - 1.5 x IQR and greater than Q3 + 1.5 x IQR where IQR = Q3-Q1, are identified as outliers.
\item Mean and Standard Deviation Method: Identifies all values less than  mean - k times standard_deviation and greater than mean + k times standard_deviation as outliers.
}}
\usage{
soc_get_oultiers(df, col, method = "SD", thresh = 3)
}
\arguments{
\item{df}{Dataframe in  which outliers are to be identified.}

\item{col}{Column vector in the dataframe based on which outliers are to be identified}

\item{method}{Name of the outlier identification method to be used.  "IQR" for IQR method and "SD" for mean and standard deviation method}

\item{thresh}{The value of k in the Mean and Standard Deviation Method formula above}
}
\value{
A dataframe which is aubset of original dataframe  containing only rows corresponding to outliers.
}
\description{
Returns outliers in the dataset based on values of a variable
This function identifies outliers in the dataset based on either of the following methods:
\enumerate{
\item Interquartile Range (IQR) Method: Identifies all values less than Q1 - 1.5 x IQR and greater than Q3 + 1.5 x IQR where IQR = Q3-Q1, are identified as outliers.
\item Mean and Standard Deviation Method: Identifies all values less than  mean - k times standard_deviation and greater than mean + k times standard_deviation as outliers.
}
}
\examples{
soc_get_oultiers(df,"Wages_Euros","SD",3)
}
