% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_identification.R
\name{soc_get_outliers}
\alias{soc_get_outliers}
\title{Find outliers in data.}
\usage{
soc_get_outliers(df, col, method = "SD", thresh = 3)
}
\arguments{
\item{df}{Dataframe in  which outliers are to be identified.}

\item{col}{Column vector in the dataframe based on which outliers are to be identified}

\item{method}{Name of the outlier identification method to be used.  "IQR" for IQR method and "SD" for mean and standard deviation method}

\item{thresh}{The value of k in the Mean and Standard Deviation Method formula above}
}
\value{
A dataframe which is subset of original dataframe containing only rows corresponding to outliers.ggpl
}
\description{
Returns outliers in the dataset based on values of a variable.
This function identifies outliers in the dataset based on either of the following methods:
\enumerate{
\item Interquartile Range (IQR) Method: Identifies all values less than Q1 - 1.5 x IQR and greater than Q3 + 1.5 x IQR where IQR = Q3-Q1, are identified as outliers.
\item Mean and Standard Deviation Method: Identifies all values less than  mean - k times standard_deviation and greater than mean + k times standard_deviation as outliers.
}
}
\examples{
library(dplyr)
small_data <- data.frame(age = - c(18, 20, 20), Wages_Euros = c(300000, 575000, 150000))
soc_get_outliers(small_data,"Wages_Euros","SD",3)
}
