% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playerranking.R
\name{rankingplayers}
\alias{rankingplayers}
\title{Ranking soccer players}
\usage{
rankingplayers(df, col_name, ...)
}
\arguments{
\item{df}{Dataframe containing players information}

\item{col_name}{A column name in the dataframe to be used to rank the players}

\item{...}{optional arguments for col_name}
}
\value{
Returns a dataframe of names representing the ranking of the players based on the specified columns
}
\description{
Given a soccer dataset, as well as column names, the function will rank the
players by their names based on the column selected by the user.
}
\examples{
rankingplayers(df,"Goals_total","Assists_Total")
}
